requirements.txt
streamlit 
langchain_groq 
python-magic
pdfplumber
docx2txt
markdown
